Split up code into files
