+++ oscovida/__init__.py
@@ -1,1 +1,6 @@
+from pandas.plotting import register_matplotlib_converters
+
+# suppress warning
+register_matplotlib_converters()
+
 __version__ = "0.1.2"
